feat: add frontchannel-logout using keycloak.js
With the old implementation the user is not logged out of the portal
since it doesn't support backchannel logout.

Now the user gets redirected to the keycloak end session endpoint which
will trigger the frontchannel-logout for the portal.

Limitation:
If there are multiple portal session (for example in two different
browsers or devices) only the one that is also used for the 2FA Helpdesk
gets logged out. This same issue also applies to backchannel logout due
to a Keycloak bug (https://github.com/keycloak/keycloak/issues/27359).

Issue univention/dev/internal/team-nubus#1238